Changed to double

+++ src/gear_place/src/robot_commander.cpp
@@ -263,7 +263,7 @@ void RobotCommander::put_down_force_cb_(const std::shared_ptr<gear_place_interfa
   Uses the force generator to put down the gear until it makes contact with the surface
   */
   double time = 1.0;
-  vector<float> initial_pose;
+  vector<double> initial_pose;
   for(int i = 0; i < 7; i++){
       initial_pose.push_back(joint_state_msg_.position[i]);
   }
